requirements.txt
affine==2.4.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
filelock==3.14.0
fiona==1.9.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
fonttools==4.51.0
fsspec==2024.5.0
gcloud==0.18.3
GDAL @ https://download.lfd.uci.edu/pythonlibs/archived/GDAL-3.4.3-cp311-cp311-win_amd64.whl
geojson==3.1.0
geopandas==0.14.4
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
kiwisolver==1.4.5
mailersend==0.5.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyparsing==3.1.2
pyproj==3.6.1
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2024.1
PyYAML==6.0.1
rasterio==1.3.10
requests==2.32.2
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
snuggs==1.4.7
sympy==1.12
tbb==2021.12.0
thop==0.1.1.post2209072238
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.18
urllib3==1.26.18
Werkzeug==3.0.3
